feat: Add additional expired session handling.
Session expiration was previously detected by checking for a HTTP 403 Forbidden status.
The current behavior of the API seems to redirect the client to `sso.greenchoice.nl/connect/authorize` when the session is expired.
It was not verified whether the check for a 403 is still required so without additional test assurance I decided to keep it in.